Consolidate authentication logic into auth.controller.js; remove legacy user control files
