+++ Backend/routes/user.router.js
@@ -1,16 +1,21 @@
 const express = require("express");
-const { login } = require("../controller/user.control.login.js");
-const { signup } = require("../controller/user.control.signup.js");
-const { confirmEmail } = require("../controller/user.control.confirmEmail.js");
+
+const {
+  signup,
+  login,
+  confirmEmail,
+} = require("../controller/auth.controller.js");
 const { getUserPosts } = require("../controller/user.control.getMyPost.js");
-const { getOtherUserProfile } = require("../controller/otherUser.controller.js");
+const {
+  getOtherUserProfile,
+} = require("../controller/otherUser.controller.js");
 
 const router = express.Router();
 
 router.post("/signup", signup);
 router.post("/login", login);
 router.patch("/confirm-email", confirmEmail);
-router.get("/:id" , getOtherUserProfile );
-router.get("/posts/:id", getUserPosts );
+router.get("/:id", getOtherUserProfile);
+router.get("/posts/:id", getUserPosts);
 
-module.exports = router;   
+module.exports = router;
